refactoring the structure of the project according to the given comments - separating one big controller/model into the smaller parts
