+++ controllers/articles.controllers.js
@@ -1,14 +1,4 @@
-const { selectTopics, selectArticleById, updateArticleById, selectUsers } = require('../models/news.models.js');
-
-exports.getTopics = (_, response, next) => {
-  selectTopics()
-    .then((topics) => {
-      return response.status(200).send(topics);
-    })
-    .catch((error) => {
-      next(error);
-    });
-};
+const { selectArticleById, updateArticleById } = require('../models/articles.models');
 
 exports.getArticleById = (request, response, next) => {
   selectArticleById(request.params.article_id)
@@ -26,16 +16,6 @@ exports.patchArticleById = (request, response, next) => {
       const timeOffset = updatedArticle.created_at.getTimezoneOffset() * 60000;
       updatedArticle.created_at = updatedArticle.created_at.getTime() - timeOffset;
       response.status(201).send(updatedArticle);
-    })
-    .catch((error) => {
-      next(error);
-    });
-};
-
-exports.getUsers = (_, response, next) => {
-  selectUsers()
-    .then((users) => {
-      return response.status(200).send(users);
     })
     .catch((error) => {
       next(error);
