+++ models/topics.models.js
@@ -1,0 +1,8 @@
+const db = require('../db/connection.js');
+
+exports.selectTopics = () => {
+  return db.query('SELECT * FROM topics')
+    .then(({ rows }) => {
+      return rows;
+    });
+};
