+++ models/users.models.js
@@ -1,0 +1,8 @@
+const db = require('../db/connection.js')
+
+exports.selectUsers = () => {
+  return db.query('SELECT * FROM users')
+    .then(({ rows }) => {
+      return rows;
+    });
+};
